// src/models/SalesItem.ts
export class SalesItem {
  $id: string;
  SalesItemId: number;
  POSItemId: number;
  ExternalReferenceId: any;
  Name: string;
  DisplayName: string;
  Description: string;
  CategoryName: string;
  GroupDetailName: string;
  ShortDescription: string;
  Price: number;
  PriceChanges : any;
  ItemOptionSetId: 50078;
  BaseImageName: string;
  ModifierGroups: any;
  DefaultOptions: any;
  IsVisible: boolean;
  IncludedModifierCount: number;
  DiscountPrice: any;
  FreeModifierCount: any;
  DiscountName: any;
  CaloricValue: number;
  IsAvailable: boolean;
  ModGroups: any;
}